im havin issues
