+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/PowerPlay23_24/AutoTemplate.java
@@ -9,7 +9,6 @@ import com.qualcomm.robotcore.hardware.Servo;
 
 
 @Disabled
-@Autonomous
 public class AutoTemplate extends LinearOpMode {
     private Servo lc;
     private Servo rc;
@@ -26,7 +25,7 @@ public class AutoTemplate extends LinearOpMode {
 
     private long mulitplier = 500;
     @Override
-    public void runOpMode() throws InterruptedException {
+    public void runOpMode(){
         waitForStart();
         if (opModeIsActive()) {
             // Put run blocks here.
@@ -49,10 +48,10 @@ public class AutoTemplate extends LinearOpMode {
 
     }
     private void initialise(){
-        frontRight = hardwareMap.get(DcMotorEx.class, "frontRight");
-        backRight  = hardwareMap.get(DcMotorEx.class, "backRight");
-        frontLeft  = hardwareMap.get(DcMotorEx.class, "frontLeft");
-        backLeft   = hardwareMap.get(DcMotorEx.class, "backLeft");
+//        frontRight = hardwareMap.get(DcMotorEx.class, "frontRight");
+//        backRight  = hardwareMap.get(DcMotorEx.class, "backRight");
+//        frontLeft  = hardwareMap.get(DcMotorEx.class, "frontLeft");
+//        backLeft   = hardwareMap.get(DcMotorEx.class, "backLeft");
 
         frontRight.setDirection(DcMotor.Direction.FORWARD);
         backRight.setDirection(DcMotor.Direction.FORWARD);
@@ -63,11 +62,11 @@ public class AutoTemplate extends LinearOpMode {
         frontLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
         backLeft.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
 // janx.armInit("clawLeft","clawRight","nodder","armExtension","arm rotations");
-        ext =       hardwareMap.get(DcMotorEx.class,"armExtension");
-        turn =      hardwareMap.get(DcMotorEx.class,"arm rotations");
-        nodder =    hardwareMap.get(Servo.class,"nodder");
-        lc  =       hardwareMap.get(Servo.class, "clawLeft");
-        rc  =       hardwareMap.get(Servo.class,"clawRight");
+//        ext =       hardwareMap.get(DcMotorEx.class,"armExtension");
+//        turn =      hardwareMap.get(DcMotorEx.class,"arm rotations");
+//        nodder =    hardwareMap.get(Servo.class,"nodder");
+//        lc  =       hardwareMap.get(Servo.class, "clawLeft");
+//        rc  =       hardwareMap.get(Servo.class,"clawRight");
 
         ext.setDirection(DcMotor.Direction.FORWARD);
         ext.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
