+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Template.java
@@ -7,7 +7,7 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 
-import org.firstinspires.ftc.teamcode.PowerPlay23_24.TemplateJanx;
+import org.firstinspires.ftc.teamcode.TemplateJanx;
 
 @Disabled
 @TeleOp(name = "what will show up on driver hub")
